Adding misc network tests.

+++ check/magma/servers/http/http_check.h
@@ -9,7 +9,7 @@
 
 /// http_check_network.c
 bool_t check_http_read_to_empty(client_t *client);
-size_t check_http_content_length_get(client_t *client);
+int32_t check_http_content_length_get(client_t *client);
 bool_t check_http_content_length_test(client_t *client, uint32_t content_length, stringer_t *errmsg);
 bool_t check_http_network_basic_sthread(stringer_t *errmsg, uint32_t port, bool_t secure);
 bool_t check_http_network_options_sthread(stringer_t *errmsg, uint32_t port, bool_t secure);
